seperate it
